requirements.txt
transformers
sentencepiece
PyPDF2 
transformers
python-docx
reportlab
docx2txt
pdfkit
streamlit
torch
